requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
dnspython==2.3.0
exceptiongroup==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
outcome==1.2.0
PySocks==1.7.1
python-engineio==4.4.1
python-socketio==5.8.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.6
wsproto==1.2.0
yarl==1.9.2
